DataService/requirements.txt
Django==2.2.11
django-filter==2.2.0
django-random-queryset==0.1.3
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
inflection==0.3.1
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.1
